Make momos recipe page and indentation
The moms recipe page contains the basic boilerplate on top of that is the heading in h1 folowed by an image. It contains a description of two paraghraphs. It has an unordered list of ingredients and an ordered list of steps to be followed. Also added indentation to every file to make the code readable and to create a parent-child relation